+++ lib/rigor/inference/method_dispatcher/constant_folding.rb
@@ -17,6 +17,7 @@ require_relative "../builtins/pathname_catalog"
 require_relative "../builtins/random_catalog"
 require_relative "../builtins/struct_catalog"
 require_relative "../builtins/encoding_catalog"
+require_relative "../builtins/re_catalog"
 
 module Rigor
   module Inference
@@ -1083,7 +1084,9 @@ module Rigor
           [Pathname, [Builtins::PATHNAME_CATALOG, "Pathname"]],
           [Random, [Builtins::RANDOM_CATALOG, "Random"]],
           [Struct, [Builtins::STRUCT_CATALOG, "Struct"]],
-          [Encoding, [Builtins::ENCODING_CATALOG, "Encoding"]]
+          [Encoding, [Builtins::ENCODING_CATALOG, "Encoding"]],
+          [Regexp, [Builtins::REGEXP_CATALOG, "Regexp"]],
+          [MatchData, [Builtins::REGEXP_CATALOG, "MatchData"]]
         ].freeze
         private_constant :CATALOG_BY_CLASS